pyproject.toml
[tool.poetry]
name = "referral-system"
version = "0.1.0"
description = ""
authors = ["Henish Patel <200120107182@git.org.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
django = "^5.0.4"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^69.2.0"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
